refactor route
